producer/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
gunicorn==23.0.0
kafka-python